Working on refactor
