+++ src/fastapi_aad_auth/_base/authenticators/__init__.py
@@ -1,0 +1,1 @@
+from fastapi_aad_auth._base.authenticators.session import SessionAuthenticator
