+++ src/fastapi_aad_auth/_base/state.py
