+++ src/fastapi_aad_auth/config.py
@@ -6,46 +6,8 @@ import uuid
 from pkg_resources import resource_filename
 from pydantic import BaseSettings, DirectoryPath, Field, FilePath, HttpUrl, SecretStr, validator
 
-
-def bool_from_env(env_value):
-    """Convert environment variable to boolean."""
-    if isinstance(env_value, str):
-        env_value = env_value.lower() in ['true', '1']
-    return env_value
-
-
-def list_from_env(env_value):
-    """Convert environment variable to list."""
-    if isinstance(env_value, str):
-        env_value = [u for u in env_value.split(',') if u]
-    return env_value
-
-
-def expand_doc(klass):
-    """Expand pydantic model documentation to enable autodoc."""
-    docs = ['', '', 'Keyword Args:']
-    for name, field in klass.__fields__.items():
-        default_str = ''
-        if field.default:
-            default_str = f' [default: ``{field.default}``]'
-        module = field.outer_type_.__module__
-        if module != 'builtins':
-            if hasattr(field.outer_type_, '__origin__'):
-                type_ = f' ({field.outer_type_.__origin__.__name__}) '
-            elif not hasattr(field.outer_type_, '__name__'):
-                type_ = ''
-            else:
-                type_ = f' ({module}.{field.outer_type_.__name__}) '
-        else:
-            type_ = f' ({field.outer_type_.__name__}) '
-        env_var = ''
-        if 'env' in field.field_info.extra:
-            env_var = f' (Can be set by ``{field.field_info.extra["env"]}`` environment variable)'
-        docs.append(f'    {name}{type_}: {field.field_info.description}{default_str}{env_var}')
-    if klass.__doc__ is None:
-        klass.__doc__ = ''
-    klass.__doc__ += '\n'.join(docs)
-    return klass
+from fastapi_aad_auth.utilities import bool_from_env, expand_doc, list_from_env
+from fastapi_aad_auth.providers.aad import AADConfig
 
 
 @expand_doc
@@ -59,8 +21,7 @@ class RoutingConfig(BaseSettings):
     page (defaults to the application root), and the ``post_logout_path`` for any specific routing
     once a logout has completed.
     """
-    login_path: str = Field('/login/oauth', description="Path for initiating the AAD oauth call", env='FASTAPI_AUTH_LOGIN_ROUTE')
-    login_redirect_path: str = Field('/login/oauth/redirect', description="Path for handling the AAD redirect call", env='FASTAPI_AUTH_LOGIN_REDIRECT_ROUTE')
+    oauth_base_route: str = Field('/oauth', description="Base Path for initiating the oauth calls", env='FASTAPI_OAUTH_BASE_ROUTE')
     logout_path: str = Field('/logout', description="Path for processing a logout request", env='FASTAPI_AUTH_LOGOUT_ROUTE')
     landing_path: str = Field('/login', description="Path for the login UI page", env='FASTAPI_AUTH_LOGIN_UI_ROUTE')
     user_path: Optional[str] = Field('/me', description="Path for getting the user view", env='FASTAPI_AUTH_USER_ROUTE')
@@ -109,40 +70,6 @@ class LoginUIConfig(BaseSettings):
 
 
 @expand_doc
-class AADConfig(BaseSettings):
-    """Configuration for the AAD application.
-
-    Includes expected claims, application registration, etc.
-
-    Can also provide additional client application ids to accept.
-
-    A list of roles can be provided to accept (requires configuring the
-    roles in the AAD application registration manifest)
-    """
-    client_id: SecretStr = Field(..., description="Application Registration Client ID", env='AAD_CLIENT_ID')
-    tenant_id: SecretStr = Field(..., description="Application Registration Tenant ID", env='AAD_TENANT_ID')
-    client_secret: Optional[SecretStr] = Field(None, description="Application Registration Client Secret (if required)", env='AAD_CLIENT_SECRET')
-    scopes: List[str] = Field(["Read"], description="Additional scopes requested")
-    client_app_ids: Optional[List[str]] = Field(None, description="Additional Client App IDs to accept tokens from (when running as a backend service)",
-                                                env='AAD_CLIENT_APP_IDS')
-    strict: bool = Field(True, description="Check that all claims are provided", env='AAD_STRICT_CLAIM_CHECK')
-    api_audience: Optional[str] = Field(None, description="Corresponds to the Application ID URI - used for token validation, defaults to api://{client_id}",
-                                        env='AAD_API_AUDIENCE')
-    redirect_uri: Optional[HttpUrl] = Field(None, description="The redirect URI to use - overwrites the default path handling etc",
-                                            env='AAD_REDIRECT_URI')
-    prompt: Optional[str] = Field(None, description="AAD prompt to request", env='AAD_PROMPT')
-    domain_hint: Optional[str] = Field(None, description="AAD domain hint", env='AAD_DOMAIN_HINT')
-    roles: Optional[List[str]] = Field(None, description="AAD roles required in claims", env='AAD_ROLES')
-
-    class Config:  # noqa D106
-        env_file = '.env'
-
-    _validate_strict = validator('strict', allow_reuse=True)(bool_from_env)
-    _validate_client_app_ids = validator('client_app_ids', allow_reuse=True)(list_from_env)
-    _validate_roles = validator('roles', allow_reuse=True)(list_from_env)
-
-
-@expand_doc
 class AuthSessionConfig(BaseSettings):
     """Authentication Session configuration.
 
@@ -188,22 +115,22 @@ class Config(BaseSettings):
     """The overall configuration for the AAD authentication."""
 
     enabled: bool = Field(True, description="Enable authentication", env='FASTAPI_AUTH_ENABLED')
-    aad: AADConfig = Field(None, description="The AAD configuration to use")
+    providers: List[BaseSettings] = Field(None, description="The provider configurations to use")
     auth_session: AuthSessionConfig = Field(None, description="The configuration for encoding the authentication information in the session")
     routing: RoutingConfig = Field(None, description="Configuration for routing")
     session: SessionConfig = Field(None, description="Configuration for the session middleware")
     login_ui: LoginUIConfig = Field(None, description="Login UI Configuration")
-    user_klass: type = Field('fastapi_aad_auth.oauth.state:User',
+    user_klass: type = Field('fastapi_aad_auth._base.state:User',
                              description="User class to use within the AADOAuthBackend, this will be treated as an import path "
                              "if provided as a string, with the last part the class to load", env='FASTAPI_AUTH_USER_KLASS')
 
     class Config:  # noqa D106
         env_file = '.env'
 
-    @validator('aad')
-    def _validate_aad(cls, value):
+    @validator('providers')
+    def _validate_providers(cls, value):
         if value is None:
-            value = AADConfig(_env_file=cls.Config.env_file)
+            value = [AADConfig(_env_file=cls.Config.env_file)]
         return value
 
     @validator('auth_session')
