+++ src/fastapi_aad_auth/logging.py
@@ -1,0 +1,1 @@
+from logging import getLogger
