+++ src/fastapi_aad_auth/mixins/__init__.py
@@ -1,0 +1,2 @@
+"""Mixins for use in objects"""
+from fastapi_aad_auth.mixins.logging import LoggingMixin  # noqa F401
