--- src/fastapi_aad_auth/oauth/__init__.py
@@ -1,4 +1,0 @@
-"""Handlers for oauth."""
-
-from fastapi_aad_auth.oauth.aad import AADOAuthBackend  # noqa F401
-from fastapi_aad_auth.oauth.state import AuthenticationState  # noqa F401
