--- src/fastapi_aad_auth/oauth/authenticators/__init__.py
@@ -1,1 +1,0 @@
-from fastapi_aad_auth.oauth.authenticators.session import AADSessionAuthenticator  # noqa F401
