--- src/fastapi_aad_auth/oauth/validators/__init__.py
@@ -1,3 +1,0 @@
-"""Validators for different authentication methods."""
-from fastapi_aad_auth.oauth.validators.session import SessionValidator as AADSessionValidator, get_session_serializer  # noqa F401
-from fastapi_aad_auth.oauth.validators.token import AADTokenValidator  # noqa F401
