+++ src/fastapi_aad_auth/providers/__init__.py
@@ -1,0 +1,3 @@
+"""Handlers for oauth."""
+
+from fastapi_aad_auth.providers.aad import AADConfig, AADProvider  # noqa F401
