+++ src/fastapi_aad_auth/providers/ms-logo.png
